recruitment-agency-app/requirements.txt
langgraph
langchain-groq
langchain-core
PyPDF2
pydantic
python-dotenv
gradio